feat(linter): rename example rule
