+++ oak-terraform-linter/tests/integration/configs/naming-strict.json
@@ -1,7 +1,7 @@
 {
   "rules": [
     {
-      "ruleType": "naming-convention",
+      "ruleType": "resource-naming",
       "enabled": true,
       "params": {
         "pattern": "^[a-z0-9_]+$"
